requirements.txt
#requirements.txt

Django~=3.2.10
djangorestframework